+++ example/lib/pages/home/home.dart
@@ -2,7 +2,7 @@ import 'package:get_server/get_server.dart';
 
 class HomePage extends GetView {
   @override
-  build(Context context) {
+  Widget build(BuildContext context) {
     return Text('Hello, you are on home');
   }
 }