+++ example/lib/pages/html/html.dart
@@ -4,7 +4,7 @@ import 'package:get_server/get_server.dart';
 
 class HtmlPage extends GetView {
   @override
-  build(Context context) {
+  build(BuildContext context) {
     final path = '${Directory.current.path}/example/web/index.html';
     return Html(path);
   }