+++ example/lib/pages/json/json.dart
@@ -2,7 +2,7 @@ import 'package:get_server/get_server.dart';
 
 class JsonPage extends GetView {
   @override
-  build(Context context) {
+  Widget build(BuildContext context) {
     return Json({
       "fruits": ["banana", "apple", "orange"]
     });