+++ example/lib/pages/upload/upload.dart
@@ -4,7 +4,7 @@ import 'package:get_server/get_server.dart';
 
 class UploadPage extends GetView {
   @override
-  Future<Widget> build(Context context) async {
+  Future<Widget> build(BuildContext context) async {
     final upload = await context.file('file');
     final data = {
       "nameFile": upload.name,